Updated translations
Some more german translations and core rulebook page informations.

To make the translation for the positive quality "Resistence for
pathogens and toxins" visible, I had to update the qualities.xml file.
It did not contain an id tag for that quality.

That change works for me, but I'm not 100% sure if it is correct for all
situations. Please check.