Weapon & Weapon Accessory Fixes
Fichetti Security 600's integral stock has been changed to an integral
folding stock, as per SR5 p. 426.

Renamed the existing Gecko Grip entry to "Gecko Grip (for Weapon with No
Stock Slot)" and added a new accessory called "Gecko Grip (for Weapon
with Stock Slot)" to indicate that the Gecko Grip accessory takes up the
weapon's Stock slot if it has one, as per R&G p. 52.

Corrected the reference for Vintage to GH3 p. 3.

Corrected Ammo Skip modification so that it takes up an Underbarrel
slot, as per the table on HT p. 183.

Corrected Chameleon Coating modification so that it takes up a Side
slot, as per the table on HT p. 183.

Corrected Electronic Firing modification so that it takes up a Barrel
slot, as per the table on HT p. 183.

Corrected Mounted Crossbow and Underbarral Laser modifications so that
the former occupies both the Barrel and Underbarrel slots and the latter
occupies both the Side and Underbarrel slots, as per the table on HT p.
182. This required a new tag to be added an the UI functions for
showing, filtering, and selecting weapons accessories to be updated,
hopefully this works fine.

Added a formal entry to the Melee Contacts of the Defiance EX-Shocker
that contains the information from the weapon's description. Where
information was missing, the entry's properties have been filled in to
be as consistent as possible with Stun Baton (10 charges, -5 AP) and the
Clubs skill for using pistols as melee weapons.

Altered Tasers to have Light Pistol concealability. This is different
from 4e, so here are the justifications for the difference based on RAW:
1. According to SR5 p. 432, Hidden Gun Arm Slide "can fit a hold-out,
light pistol, or taser". This means a taser must be smaller than a heavy
pistol, since this accessory cannot fit heavy pistols or machine
pistols, and since size directly relates to concealability, indicates
that a taser must have a more negative concealability than a heavy
pistol or a machine pistol.
2. According to the end of Cybergun Modifications section of CF p. 90,
tasers are treated as light pistols for the purposes of implantation.
Implanted cyberweapons' essence and capacity costs relate directly to
their size, with the capacity cost for all cyberguns except grenade
launchers being equal to (6 + concealability). Since a cyberimplanted
taser would have the same essence and capacity cost as a cyberimplanted
light pistol, which is less than those of a cyberimplanted heavy pistol,
this would indicate that a taser has the same concealability as a light
pistol.
3. The Defiance EX-Shocker contains contacts that can be used as a melee
weapon, as written on SR5 p. 424. The Reach stat of the taser using
these contacts is 0. This means they have the same Reach as a Sap or
Knife, both of which have a concealability of -4, the same
concealability as a light pistol. A Club has the same concealability as
a heavy pistol and has a Reach stat of 1. Assuming the contacts are
located at the front end of the weapon, a taser with a size and
concealability of a heavy pistol would would have a Reach of 1 with its
contacts, which is not the case.