Drone, Vehicle, Drone Mod, and Vehicle Mod Fixes
Added a pair of drone arms and drone legs to the Ares Duelyst because it
is an anthropomorphic drone.

Added a pair of drone arms to the Transys OfficeMaid drone, as per
references in the fluff text on SS p. 180 to "its robotic arms".

Added a pair of drone arms and drone legs to the Modified Renraku
Manservant-3 and recategorized it as Anthropomorphic, as per the
alternate models of the Aztechnology Criado Juan on R5 p. 145 listing
"Renraku Manservant" and its description as an anthropomorphic drone.

Added a custom entry to the Dustoff drone for its armored Valkyrie
Module. Game-wise, this should behave like a combination of Rigger
Cocoon and Valkyrie Module, as per BB p. 23.

Added a pair of drone arms and drone legs to the Caduceus CAD drone and
the Snake Fingers cyberware mod, as per BB p. 23 (Walker mod means it
has two legs). Due to it having a pair of arms and legs, the drone was
recategorized to anthropomorphic.

Added a pair of synthetic drone arms and synthetic drone legs to the
Mitsuhama Akiyama, as per HT p. 190.

Added a pair of drone arms and drone legs and a rating 1 tutorsofts for
Blades, Clubs, and Unarmed Combat to the Sparring Drone, as per HT p.
190.

Added the following missing weapon classes from the SR5 weapon mounts:
Blades, Clubs, Exotic Melee Weapons, Crossbows, and Sporting Rifles for
Standard Mounts, and Blades, Clubs, Exotic Melee Weapons, Crossbows, and
Exotic Ranged Weapons for Heavy Mounts. Also removed Sniper Rifle from
Standard Mounts because it has a more positive concealability than an
Assault Rifle.

Added a bodymax tag to Gecko Tips (Bod 4-6) so that it's not available
to vehicles with a body score higher than 6.

Removed the Melee, Machine Pistols, and Submachine Guns categories from
Rigger 5.0 Standard and Heavy Weapon Mounts, but added a note to add
them back in as soon as errata is released.

Fixed the Winch (Enhanced) vehicle mod requiring 1 body mod slot instead
of 2, as per the table on R5 p. 167.

Added two new items for add-ons to drone weapon mounts: blow-away panels
and pop-out mods. Pop-out mod's rating is meant to represent the
modified MP of the weapon mount.

Added the ability to add Bodyware to drone arms, drone legs, synthetic
drone arms, and synthetic drone legs. While most Bodyware would be
useless, there are some pieces that could be useful (e.g. Foot Anchor,
Magnetic System, Retractable Climbing Claws).

Added new entries for the Daihatsu-Caterpillar Horseman ModPods, as per
R5 p. 42. Since each one decreases acceleration by 1 and the Horseman
only has an acceleration of 2 to begin with, the cap of one modpod max
is soft-enforced by the fact that adding more than one modpod would drop
the Horseman's acceleration to 0.