Fixed Multiple Mount Slot Weapon Accessories
A new drop-down menu was added for weapon accessories that require
multiple accessory slots. The ExtraMount field is optional, but
otherwise works almost identical to the Mount field. When selecting an
accessory that requires multiple slots and with not enough available
slots, at least one of the Mount fields will always automatically switch
to "None" to indicate that it's missing a mount slot.