Weapon Fixes, Additions
Added an entry for the Shiawase Arms Simoom from Hard Targets. The main
entry is in Specialty Armor, but it adds an item to the weapons list
when taken.

Fixed the bug where Riot Shield and Ballistic Shield were actually
charging twice their cost due to the player being charged once for the
armor entry and a second time for the weapon entry. The weapon entry is
now free, but is hidden from the menu and can only be added by taking
the armor entry.

Hid weapon entries for cyberweapons and gear items (e.g. Grapple Gun,
Micro-Flare Launcher) so that they cannot be taken (for free) without
buying the entry they are tied to first.

Added missing Stock and Slide weapon accessory slots to all tasers,
light pistols and heavy pistols, as well as a few SMGs and a sniper
rifle that was missing those slots. While it unclear if Tasers, Light
Pistols, and Heavy Pistols should have those mod slots in RAW, they do
have those slots in HeroLab's Shadowrun module, which is an official
product of CGL, so it can be treated like RAW.

Added missing Top and Underbarrel weapon accessory slots to the Ares
Armatus.