Weapon Accessory Gear Fix, Smartgun + Imaging Scope
Weapon accessories can now add gear when <gears> tags are specified (in
addition to weapons themselves also being able to add gear).

Added built-in micro camera and vision magnification to Imaging Scope.

Changed smartgun accessories to include a laser range finder and a
micro-camera instead of allowing for the addition of vision enhancement
gear (this is now done through adding said gear as a plugin to the micro
camera).