Vehicle Mod, Vehicle Stat Changes
Application Changes:

- Added the ability to distinguish between primary and offroad speed and
acceleration, in addition to primary and offroad handling.
- Added the ability for vehicle mods to alter seat count.
- Added the ability for vehicle mods that alter speed, handling, and/or
acceleration to properly subtract values and also to support percentage
modifications instead of just flat ones.

Data Changes:

- Added vehicle entries for the Cocotaxi and the Camellos from Hard
Targets and the Ares Garuda and the Evo Aquavida 2 from Rigger 5.0.
- Added stats for secondary acceleration and speed (i.e. offroad
acceleration and speed for ground vehicles, alternate locomotion for all
others) where they were missing, e.g. from the Evo Falcon-EX, the
Corsair Trident, and the Evo Krokodil.
- Corrected the seats for all vehicles that have separate front/back
sections so that Chummer always displays the total seat count, not just
the seat count for the larger section.
- Corrected the availability of most vehicles from Stolen Souls (to 4)
and the Ares Roadmaster (to 8).