Advanced Programs (+ Misc) Part 5
Removed the duplicate entry for Chatty.

MainController is now a child of CharacterShared.

Added the ability for the cyberware tab to be removed via the disabletab
bonus.

Removed the ability for AIs to get any sort of 'ware besides essence
holes and essence antiholes.

Fixed Inherent Program allowing for the selection of programs other than
Common Programs and Hacking Programs.

Fixed the base Shadowrun sheet type not properly displaying the
requirements of advanced programs.