Updates german character sheets
Updates the german character sheets to reflect some of the layout
changes done in the english version.

List of changes:
- Updated skills table to match the english layout (display attribute
name, column widths).
- Moved "class=tableborder" of LimitsBlock to inner table. Before this,
the "hseparator" after that block is inside of the table border. No the
layout matches other blocks.
- Increased contact notes column span to entire table. Before this, it
only covers 4 of the 6 columns.
- Translated ranged weapon range table abbreviation from "S" for short
to "K" for "Kurz", which is the term already used in the UI weapon pane.
- Translated vehicle table header for "Seats" to "Sitze".