Small character sheet fixes
A set of small fixes on the character sheets.

- Expenses (karma and nuyen) sorting removed from xslt. Due to date
format (mm/dd/yyyy) that sort destroyed the ordering of the entries.
Sorting is already done in the program.

- Swapped order of "ranges" and "accessories" fragments for under barrel
weapons of ranged weapons. Now its order matches the order of that
fragments for the weapon itself.

- Move "TableTitle" of the divs for notes,  background, concept and
description from before the div into the div. Otherwise the table title
remains visible if the block is hidden via the javascript "show" option.

- Left aligned header for physical and stun damage track. So far, only
the first line of the physical track was left aligned. That looked
strange in german (the second line is longer than the first one).