Small character sheet fixes (#1583)
* Small character sheet fixes

A set of small fixes on the character sheets.

- Expenses (karma and nuyen) sorting removed from xslt. Due to date
format (mm/dd/yyyy) that sort destroyed the ordering of the entries.
Sorting is already done in the program.

- Swapped order of "ranges" and "accessories" fragments for under barrel
weapons of ranged weapons. Now its order matches the order of that
fragments for the weapon itself.

- Move "TableTitle" of the divs for notes,  background, concept and
description from before the div into the div. Otherwise the table title
remains visible if the block is hidden via the javascript "show" option.

- Left aligned header for physical and stun damage track. So far, only
the first line of the physical track was left aligned. That looked
strange in german (the second line is longer than the first one).

* Fixed page references to german core rulebook

Fixed page references for critters, spirits and sprites.
Fixed name translation for Devil Rat.

* Revert "Fixed page references to german core rulebook"

This reverts commit 04a9bedb6f8eaae0bc561aa459acf0f29e003760.

* Small character sheet fixes

Two small fixes on the game master summary sheet.
- Fixed attribute title for edge. Was ESS instead of EDG.
- Added intended spaces around rating of cyberware. Replaced a simple
space character by <xsl:text> </xsl:text> to retain the space in the
output.