Housecleaning
Updated AWSSDK Packages for ChummerDataViewer.

Marked all projects in the solution as needing to be CLSCompliant. Modified certain classes, modified certain class member visibilities, and added new getters to reduce CLS non-compliant sections.

Moved the entirety of clsMainController over to CommonFunctions and marked all of the methods as static like they should be. This is a stopgap until all static methods that rely on a Character argument get moved over to become methods in the main clsCharacter class.

Greatly simplified the methods that alter a character's home node to rely on the property's doubly-linked nature instead of browsing through entire lists.

Added the translator to the Release config to assist with profiling and static code analysis.