Housecleaning Part 2
Slightly increased performance of BindingListDisplay redraws when the underlying BindingList is changed.

Fixed a bug with KnowledgeSkillControl that caused it to not properly update type and specialization values when the selected skill is changed.

Added some previously unhandled improvements to clsCharacter's improvement hook lists, and made them both HashSets to increase the performance of ImprovementHook calls.

Fixed a regression where changing skill values would sometimes not properly update the displayed value.