String-focused Optimization/Standardization Sweep
Standardized the parsing of fields starting with "FixedValues" and "Variable".

Replaced all instances of string::Contains() and string::IndexOf() that checked for single-character strings with versions that just check for the char.

Added four extension methods to the string class that allow for faster trimming of strings and faster deletion of characters compared to string::Replace(). Wherever applicable, Replace() calls have been replaced with one of these four methods.

Replaced Replace() calls that were called to remove the availability type indicators at the end of availability strings with significantly faster Substring() calls.