Eliminated Unnecessary ScheduleCharacterUpdate() Calls, Crash Fix
Fixed a crash related to legacy cyberware items.

Eliminated ScheduleCharacterUpdate() calls in places where they completely unnecessary, and added in handling of _blnSkipRefresh in a few methods that previously did not handle it but should have (e.g. Weapon Accessory-related checkbox events). This should markedly increase Chummer5 performance when just browsing around a character with it, not making any changes.