Housekeeping, Regression Fixes
Fixed a regression where a character's current essence score was added to the essence attribute's base. This was originally meant as some sort of fix to printing essence on sheets, but since that's now handled differently anyway, the line was commented out.

Fixed a regression where Visual Studio's form designer could not load in frmCareer or frmCreate.

BindingListDisplay now always "loads fast" (i.e. it loads immediately when constructed), and controls that use them now load alongside all other items in frmCreate and frmCareer instead of only when their tab is clicked and they become visible.

Mugshot lists are now stored in memory as Image lists and are converted to and from Base64 strings when they are saved or loaded, rather than being saved and loaded as Base64 strings and converted into images whenever they are displayed. This should make the controls that displays mugshots much more responsive, especially if they contain large images, at the cost of slowing down character saving and character loading (printing is unaffected because it would need to export both the string and the image anyway), and slightly increasing memory footprint.

Increased drawing performance of any new mugshots loaded from file formats that do not use a 32bppPArgb pixel format by explicitly converting to that format when the mugshot is added.