Backend for Metamagic- and Echo-Granting Improvements
Added backend support for bonuses that add metamagics and/or echoes, as well as ones that allow the user to choose a metamagic or echo. As a side effect, the Initiation tab is enabled in Create mode as well now, but if initiations are disabled at character generation, the button to add an initiate grade is disabled.

Fixed a regression where a character would not fail the validity check at generation if they had "None" selected as their Tradition or Stream.