String Optimizations, Possible frmSelectSpell Fix
Added in a CheapReplace() method that works similar to string::Replace(), but it only generates the string with which to replace a pattern if the string contains the pattern. This makes it most suitable to replaces that would use expensive methods to generate the new value, where that expense is wasted if the base string doesn't contain any instances of the pattern.

Removed some unnecessary bits of code from frmSelectSpell. One of these removals may end up fixing a sporadically reported bug where it does not appear in Create mode.