Localization, Sheet Fixes & Enhancements
Altered the localization format so that language names comply with the formatting of CultureInfo.Name. New names are composed of two, two-letter codes separated by a hyphen, with the first code to abide by ISO 639-1 and the second code to abide by ISO 3166 formatted as lowercase.

Fixed incorrect references to certain .xslt files. Fixes #2071.

Application Changes:

- Character sheets are now printed with the number formatting of the language in which they are printed instead of with the number formatting of the language of the user's current operating system.
- Fixed a bug related to incorrect printing of certain numbers due to differences in decimal separators. Fixes #2069.