Added weapon alternative range to sheet
- German translations for new sheet texts (natural recovery).
- Added alternate weapon ranges for ranged weapons.
Might not be the best implementation in terms of "clean xslt and best
practive" but works.
Implemented using a simple for-each loop over ranges and
alternateranges.
Partly solves issue #2070 by adding the alternate range table. Still
missing a name/description of the range table that needs to be added to
the print xml first.