XmlDocument-Invariant XPathNavigator Housecleaning
Unified all instances of XPathNavigator evaluations where the expression was invariant of the XmlDocument so that they all call the same XPathNavigator object instead of constantly creating their own. This should also make it easier to replace these evaluations in the future with something that is faster and/or more flexible than XPath.