SR Missions 1.4 FAQ Ruleset Corrections
Allowed Nothing Personal and Shadows in Focus Metropole content in the custom ruleset for SR Missions, as they are allowed by Missions 1.4.

Re-Allowed Counting Coup content in the custom ruleset for SR Missions, as the only content from it present in Chummer5 is the Coyote Mentor Spirit, which is explicitly allowed by Missions 1.4 even if the adventures from that book are not.