Fixed XPathNavigator::Evaluate() Intermediate Cast Regressions
Fixed a bunch of regressions related to intermediate casting of results from XPathNavigator::Evaluate(). For example, for numbers, the method always returns a double, so trying to convert it to a decimal before immediately converting it to a string or an int can cause problems due to the range and precision differences between decimal and double types. It also causes mix-ups with exceptions, as when the result of an Evaluate is a string and not a number, it should throw an InvalidCastException when attempted to be used as a number instead of something like OverflowException.