Cached Skill Group Disables, Fixed Nuyen Decimal Nud Location
The boolean for whether or not a skill group is disabled is now cached instead of needing to constanty query Improvements. This was originally a minor performance bottleneck.

Fixed the numerical up-down control for maximum amount of decimals to display for nuyen amounts overlapping with the numerical up-down control for minimum amount of decimals to display for nuyen amounts.