Minor Performance Fixes, Added Missing FamFamFam Icons
XmlNodes are now cached if LiveCustomData is not enabled. The reduction in System.Xml calls should create an overall increase in performance, even if the cache-resetting code makes Name/ID setters work slightly slower than before.

Replaced all instances of "new Guid()" with Guid.Empty.

Moved the newline character replacer of Language strings so that it is applied once when the string is stored, instead of every time a string is fetched.

Sped up the spell descriptor builder by having it use a StringBuilder in the loop instead of concating strings.

Added all icons that were missing from the FamFamFam icon set to the icons folder as available resources.