File Dialog QoL Enhancement, MRU Enhancements, CharacterRoster Fixes
If a user is changing the file/location of something (PDF of book, PDF of program, character file for contact, etc.), the dialog that asks the user to select a file will now automatically open to the current file/location if it exists.

Renamed the methods used for the MRUChanged EventListener to be more descriptive of what they do.

Added methods to remove or add a list of characters to MRU lists instead of just being able to add character at a time.

Methods that modify MRU lists will now only change registry entries if something in the list was changed, and they will only invoke the MRUChanged event if something in the list was changed and the appropriate argument in the method is not set to false. The latter is used to ensure that if two or more methods are called back-to-back that would invoke MRUChanged anyway, only the last method in the series will end up invoking it.

Application Changes:

- Fixed a crash when moving characters between the "Favorites" and "Recent" lists in the character roster.
- The "Favorites" list will now be shown in the character roster even when it is empty so that characters can immediately be moved to it.