Housecleaning
Cleaned up the Chummer\helpers directory. Its contents are now either implemented as extensions, or they have been renamed so as not to clash with classes that they inherit.

Added GNUv3 comment to the top of all extension .cs files.

Allowed the custom Label/Button extensions with a built-in ToolTip to use shared ToolTip objects and/or classes that inherit ToolTip.

Made minor optimizations to TreeView extension methods.