Nuyen Decimals Controls Cleaning, Enhancement
Renamed controls and properties related to the minimum number of decimal places shown for nuyen amounts; they are now labelled as "Minimum" instead of "Always", which should make their function clearer.

Changed the interaction between the two nuyen decimal places numerical up-downs. Instead of the Maximum Places value setting the maximum for the Minimum Places control, they both have fixed maxima, but whenever either value is changed, if the Minimum Places value is greater than the Maximum Places control's, then the Maximum Places value is set to that of the Minimum Places control's.