Relative CustomDataDirectory Paths, PDF Test Fix
Custom Data Directory paths that lead to a folder somewhere inside the Chummer start directory will now be saved as paths relative to the Chummer start directory instead of absolute paths for directories on the current user's machine. This is a stepping stone to allow portable settings profiles to save enabled folders in the customdata directory inside Chummer's base directory.

Application Changes:

- Fixed a bug where testing a PDF inside the options menu would automatically save the currently selected options, which would cause them to be applied even if the user would choose not to apply changes to options in the prompt when they exited the options menu.