ElasticComboBox Enhancement, Helper Tweaks, More WaitCursors
ButtonWithToolTip and LabelWithToolTip will now only set tooltip properties if an existing tooltip object is not defined in the constructor.

ElasticComboBox will now resize its DropDown menu when its data source is changed, rather than when it is clicked. This eliminates the annoying stutter effect, makes sure the computation is only fired once instead of every time the dropdown menu would be drawn, and groups the loading time behind the computational cost of resizing the dropdown menu with other loading times, which makes the UI more responsive.

Added code to change the cursor to WaitCursor and back to all sections of frmMain that create and open new forms. Added code to change the cursor to WaitCursor to Translator whenever there is a long list of items being loaded or a form is being created.