frmViewer Enhancements, Malformed ID Checking
frmViewer's language and sheet ComboBoxes are now also ElasticComboBoxes.

The label in frmViewer is now automatically moved based on localized text length. The minimum width for frmViewer is dynamically set based on said label's width and the width of the two boxes (whose texts can also be localized to be longer).

Altered the IDs for Infrasonic Generator, Mercedes Paladin, and Walther PP110 so that they can actually be parsed as GUIDs. Added in code to silently legacy shim items like these (i.e. items with a malformed id field) to make sure their proper SourceIDs can be re-fetched when a character is loaded.

Application Changes:

- When warning about duplicate GUIDs in data files, Chummer5a will now display the names of the items to which those duplicate IDs belong.
- In addition to checking for duplicate GUIDs, Chummer5a will now also check for ID fields that do not contain GUIDs, and if it finds any, it will warn users in a manner similar to the warning for duplicate GUIDs.

New Strings:

- Message_NonGuidIdWarning

Changed Strings:

- Message_DuplicateGuidWarning