MergeInto Function On Equals
Added the ability for BindingList::MergeInto() to run an action on an existing member of the BindingList if the comparison between the existing and new items returns zero (i.e. new "equals" existing).

When a new skill would be merged into a skill BindingList, but the comparison result is 0, then the existing skill's Base and Karma values become the maximum between the existing skill and the new skill, and any specializations that the new skill has but the existing skill  does not get added to the existing skill.

When a new skill group would be merged into a SkillGroup BindingList, but the comparison result is 0, then any skills that are part of the new skill group's skill list but not the existing one's get added to the existing skill group's list.