Translator Sanity Check Part 1, Enhancements
Part 1 of a rewrite of the Process methods in the Translator so that they're actually maintainable by sane human beings. Turns out that the rewrite also speeds up the code and makes the compiled binary smaller, so it's a win in all aspects.

Added an icon to the Translator tool.

Application Changes:

- The Translator tool can now be opened from within Chummer via Tools > Translator.

New Strings:

- Menu_Main_Translator