Translator Enhancements, Japanese Code Fix
The translator will now only enable the "Create" button if a properly formatted language code is given, and it will only enable the "Edit" button if an existing localization has been selected. It will now auto-fill the language name section based on the provided language code and warn the user when the code they provided does not comply with ISO 639-1 and ISO 3166-1. If a translation with the same code already exists when a new one would be created, the Translator will now ask the user if it wants to rebuild the existing localization or delete it and start anew.

Application Changes:

- Added a "Rebuild" button to the Translator tool. When used, it will regenerate the strings of an existing translation, removing any unused strings and adding any new ones that have been added after the document was initially created.

Data Changes:

- Corrected the code of the Japanese localization to ja-JP (from jp-JP).