Language Hotswapping Phase 3
Almost all strings (Contacts are the exception) now support hotswapping of languages, concluding Phase 3. Phase 4 will eventually be the ability to re-translate the UI whenever languages change, but it's of not-as-high importance because it doesn't interact with the way characters can be printed in languages different from the UI language.