Amend System cleaning and fixes
Fixed a potential issue with the amend system when its "operation" attribute could clash with amending nodes that are part of the "details" subsystem of the selection system (where "operation" attributes are used to specify what operation to perform against a detail).

Existing attributes will no longer be deleted under the "append" operation within the amend system.

Made it so that the default operation becomes "append" if there is no target node found under the amend system.

Cleaned unnecessary code in the amend system, including the unnecessary need for a return value or identifier tracking.