Major Housecleaning
Altered backed of the Weapon class so that it stores its cost in string form and converts it to decimal when necessary. This should help in the future if/when weapon costs need to be stored in string form (e.g. if their costs would need an XPath evaluation).

Extended the fix for armor mods not accounting for armor parent cost discounts to weapon accessories as well.

Application Changes:

- Migrated all instances of use of "yes" and "no" strings in data to use "True" and "False" instead respectively. Any custom content that relies on strings like these will need to be altered, and any items reliant on improvements that use these strings will need to be fixed with Reapply Improvements.