Housecleaning, ObservableCollection Expansion
A character's Complex Form, AI Advanced Program, Quality, and Critter Powers lists are now all ObversableCollections that are monitored in ways similar to how the Spells list is monitored.

Switched over the system for Quality TreeNodes so that their Tags are now pointers to Quality objects instead of strings of Quality objects' internal IDs.