Character Roster and MRU Fixes & Enhancements
Switched the MRU list system so that it's a special ObservableCollection that has a fixed number of maximum items and moves existing items instead of adding duplicates. Changes to this ObservableCollection are synchronized to Chummer5's registry key, including changes in ordering.

The character roster no longer always redraws the entire roster if any lists are changed. Instead, it will only redraw lists that are necessary to be redrawn.

Application Changes:

- Chummer5 will now live-update the watch folder node in the character roster if the underlying folder's contents are changed in any way.
- Both the Recent Characters and Favorite Characters lists in the character roster can now be sorted alphabetically. Part of #2074.
- Characters no longer lose their position in the Most Recently Used list if they are added to the Favorite Characters list.