Options Translation Regression Fix
Fixed a regression where translations would not get properly loaded for options. Fixes #2480.

Also reformatted options.xml by removing the `<options>` node that was between the base `<chummer>` node and the root nodes for the different types of options. This brings the file more in line with other data files' formatting.