Initial Backend for New Story Generator
Added initial backend for a new version of a story builder. This builder is able to handle multiple languages, potential choosing of persistent strings (instead of them being randomly generated all the time), and complicated macros (e.g. macros can actually generate the text that would define a new macro). It is also divorced from life modules, allowing it to potentially be used to generate stories outside of life module creation mode. Performance will need to be looked at in the future though, as the generation of persistent strings is more cumbersome than I'd want it to be (it seems a bit unavoidable right now, due to how processing a non-persistent macro can potentially end up producing a macro that calls for a persistent module).