Housecleaning, Weapon Accuracy Fixes
For performance reasons, the format in which essence should be displayed is now what is saved with CharacterOptions instead of the number of decimal places to which to round essence. The latter is now a derived property of the former, though it is cached for performance purposes.

Application Changes:

- Unhardcoded the weapon skill accuracy increase given by the Enhanced Accuracy power so that it can now be used through weaponskillaccuracy bonus nodes.
- Fixed an issue where bonus accuracy from Laser Sight (2050) would stack with smartgun accessories.

Data Changes:

- Fixed an issue where the Enhanced Accuracy power could be taken for Unarmed Combat.
- Fixed an issue where the accuracy bonus from multiple instances of the Enhanced Accuracy power taken for the same skill would stack with one another.