More TableLayoutPanel. FlowLayoutPanel Passes
Converted all screens other than those in the Street Gear tab to use TableLayoutPanels and/or FlowLayoutPanels instead of relying on MoveControls() stuff.

A side effect of this is that controls that are stored within such panels aren't affected by the weird, size-changing shenanigens of Visual Studio's WinForms Designer stuff.