Housecleaning, Groundwork for Separating Vehicles/Drones Form
Laid the groundwork for separating out the Vehicles/Drones form from frmCreate and frmCareer. A lot of methods that were previously tied to CharacterShared are no longer tied to it, and aside from Copy-Paste status, VehiclesTabUserControl should mostly be functional, though it has yet to be placed into frmCareer and frmCreate and properly tested.