More UI Tweaking, Weapon Accessory Backend Enhancement
Added backend support for Weapon Accessories saving their maximum rating. This is for use later on when numerical up-down controls in frmCreate are set up to be able to control the ratings of weapon accessories.

Removed a lot of TableLayoutPanel paddings that were probably unnecessary.

Switched more systems of buttons next to text to use FlowLayoutPanels instead of separate columns within a TableLayoutPanel.

Changed the default resolution from 1280x768 (which was previously a mishmash between 4:3's 1024x768 and 16:9's 1280x720) to 1280x720.