More UI Fiddling, Now with GroupBoxes
Switched the UI setup of the Cyberware tab to a FlowLayoutPanel containing GroupBoxes each with its own, docked TableLayoutPanel and each focusing on a singular aspect of the selected item (so e.g. there's a separate GroupBox for Matrix stats that bioware won't have). The FlowLayoutPanel will have scrollbars show up if it gets resized to something too small, while the GroupBoxes auto-size but have a minimum width set. This very much appears to be the best UI setup so far, and future fiddling will likely switch tab setups to something like this.

New Strings:

- String_Matrix
- String_Info