Added Backend for Translating Compound Strings
Added backend support for translating compound strings, i.e. strings where plaintext is mixed with calls to localized strings.

It's rather slow compared to just outright calling the functions themselves, so it is intended to be used to facilitate a system for translating things like Limit Modifier conditions, where users can specify text as just plaintext, just localized strings, or a mix of both.

Due to its expensive nature, it is not intended to replace instances where the components of a compound string are hardcoded, e.g. text in dice roller results windows.