Minor UI Tweaking, KC CCOB Data Regression Fixes
Armor Mods will no longer display an armor value when selected if they do not modify armor values.

Fixed a crash that would occur when selecting one of the armor mods that comes standard with the Pantheon CCOB. Fixes #3158.

The armor mods that come standard with the Pantheon CCOB will now be hidden from the list of armor mods to add to other pieces of armor, as no proper values for their availability, cost, or armor capacity cost are stated in KC.

Pantheon CCOB now comes standard with its Armored Shell, as it should.

Generic CCOB now properly comes with only one Micro-Hardpoint instead of three.